Use the empty string rather than `None` for blank database fields
For testing and development environments, it's not uncommon for some of
the database fields (eg password) to be blank. For example on Travis:
`mysql://travis@localhost/test_db`

Prior to this change, blank fields would result in the corresponding
Django config key being given a value of `None`, which can cause issues
with some of the DB backends, or other tooling that references the
DATABASES dict. If the fields are omitted entirely, Django defaults to
the empty string, so we should be using that instead:
https://docs.djangoproject.com/en/1.8/ref/settings/#password

A similar fix landed in dj-database-url:
https://github.com/kennethreitz/dj-database-url/pull/10

Fixes #56.